LayoutTests/inspector-obsolete/styles/styles-history-expected.txt

Tests resources panel history.


Running: testSetUp

Running: testSetResourceContentMinor
History length: 0

Running: testSetResourceContentMajor
History length: 1
Item 0:
body {
  margin: 20px;
  padding: 10px;
}

Running: testSetContentViaModelMinor
History length: 1
Item 0:
body {
  margin: 20px;
  padding: 10px;
}

Running: testSetContentViaModelMajor
History length: 2
Item 0:
body {
  margin: 20px;
  padding: 10px;
}
Item 1:
body {
  margin:30px;
  padding: 10px;
}
